backend/pyproject.toml
[tool.poetry]
name = "demo_backend"
version = "0.1.0"
description = "Demonstration of containerized FastAPI REST api with Postgres (SQLAlchemy+Alembic) and Redis (Celery+Flower)."
authors = ["Brent Stone <brent-stone@github.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
pandas = "^1.4.1"
scipy = "^1.8.0"
scikit-learn = "^1.0.2"
python-can = "^4.0.0"
can-isotp = "^1.7"
numpy = "^1.22.2"
fastapi = "^0.79.0"
uvicorn = {extras = ["standard"], version = "^0.17.5"}
gunicorn = "^20.1.0"
pydantic = {extras = ["dotenv", "email"], version = "^1.9.2"}
alembic = "^1.7.6"
SQLAlchemy = {extras = ["asyncio"], version = "^1.4.39"}
databases = {extras = ["asyncpg"], version = "^0.6.0"}
psycopg2-binary = "^2.9.3"
PyJWT = "^2.3.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
celery = "^5.2.7"
flower = "^1.2.0"
aiofiles = "^22.1.0"
requests = "^2.28.1"
redis = "^4.3.4"

[tool.poetry.dev-dependencies]
pytest = "^7.0.1"
tox = "^3.24.5"
flake8 = "^4.0.1"
black = "^22.1.0"
mypy = "^0.931"
pre-commit = "^2.17.0"
tox-poetry = "^0.4.1"
Sphinx = "^4.4.0"
tox-pyenv = "^1.1.0"
setup-cfg-fmt = "^1.20.0"
reorder-python-imports = "^2.6.0"
pre-commit-hooks = "^4.1.0"
sphinx-rtd-theme = "^1.0.0"
sphinx-autodoc-typehints = "^1.15.2"
bandit = "^1.7.1"
safety = "^1.10.3"
interrogate = "^1.5.0"
pandas-stubs = "^1.2.0"
pytest-asyncio = "^0.18.1"
asgi-lifespan = "^1.0.1"
coverage = "^6.3.2"
pytest-dotenv = "^0.5.2"
pytest-cov = "^3.0.0"
sqlalchemy-stubs = "^0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
env_files = ".test.env"
log_cli=true
log_auto_indent=true
log_level="DEBUG"

